ot-text追加、モック作成
