+++ addons/wiki/templates/edit.mako
@@ -410,6 +410,8 @@ ${parent.javascript_bottom()}
 </script>
 <script src="//${sharejs_url}/text.js"></script>
 <script src="//${sharejs_url}/share.js"></script>
+<script src=${"/static/public/js/sharedb-dist/sharedb.min.js"}></script>
+<script src=${"/static/public/js/sharedb-dist/ot-text.min.js"}></script>
 <link href="${node['mfr_url']}/static/css/mfr.css" media="all" rel="stylesheet" />
 <script src="${node['mfr_url']}/static/js/mfr.js"></script>
 <script src=${"/static/public/js/wiki-edit-page.js" | webpack_asset}></script>
