+++ package.json
@@ -67,6 +67,7 @@
     "select2": "3.5.1",
     "sharedb": "3.2.1",
     "sharedb-string-binding": "^1.0.0",
+    "rich-text": "^4.0.0",
     "style-loader": "^0.8.3",
     "treebeard": "https://github.com/CenterForOpenScience/treebeard.git#d2bb8f6f8104fcac040402727f431c26722b269e",
     "typeahead.js": "^0.10.5",
